+++ apps/upload-api/src/routes/index.ts
@@ -1,7 +1,7 @@
 import { FastifyInstance } from "fastify";
 import files from "./files/index.js";
 import metadata from "./metadata/index.js";
-import permissions from "./metadata/permissions/index.js";
+import permissions from "./permissions/index.js";
 import schedulerCallback from "./schedulerCallback.js";
 
 export default async function routes(app: FastifyInstance) {
