--- apps/upload-api/src/test/routes/metadata/permissions/utils/remofeFileSharing.test.ts
@@ -1,12 +1,0 @@
-import t from "tap";
-import removeFileSharing from "../../../../../routes/metadata/permissions/utils/removeFileSharing.js";
-import { PostgresDb } from "@fastify/postgres";
-
-t.test("removeFileSharing", async () => {
-  t.test("executes query with correct params", async (t) => {
-    const params: string[] = [];
-    const pg = { query: (...args: string[]) => params.push(...args) };
-    removeFileSharing(pg as PostgresDb, "fileId", "userId");
-    t.match(params[1], ["fileId", "userId"]);
-  });
-});
